Use u8 instead of usize where possible

+++ src/cube/corner.rs
@@ -1,7 +1,7 @@
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub(crate) struct Corner {
     pub(crate) piece: CornerPiece,
-    pub(crate) orientation: usize,
+    pub(crate) orientation: u8,
 }
 
 #[derive(Clone, Debug, PartialEq, Eq)]
@@ -17,11 +17,11 @@ pub(crate) enum CornerPiece {
 }
 
 impl Corner {
-    pub(super) fn new(piece: CornerPiece, orientation: usize) -> Self {
+    pub(super) fn new(piece: CornerPiece, orientation: u8) -> Self {
         Self { piece, orientation }
     }
 
-    pub(super) fn rotate(&mut self, rotation: usize) {
+    pub(super) fn rotate(&mut self, rotation: u8) {
         self.orientation = (self.orientation + rotation) % 3;
     }
 }
